# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Fitzwilliam Museum Collection Database API",
        "description": "Api for the Fitzwilliam Museum Collection Database",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/periods": {
            "get": {
                "tags": [
                    "Terminology"
                ],
                "summary": "Retrieve agents used in the database",
                "description": "A list of periods used in the database, with pagination.",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "string"
                        },
                        "example": "Roman"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "description": "Size of the page response",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "20"
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sort direction",
                        "required": false,
                        "schema": {
                            "type": "enum",
                            "enum": [
                                "asc",
                                "desc"
                            ],
                            "nullable": true
                        },
                        "example": "asc"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The request completed successfully."
                    },
                    "400": {
                        "description": "The request cannot be processed"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/api/v1/periods/{period}": {
            "get": {
                "tags": [
                    "Terminology"
                ],
                "summary": "Retrieve a period",
                "description": "A period's representation as used in the database.",
                "parameters": [
                    {
                        "name": "period",
                        "in": "path",
                        "description": "Query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "string"
                        },
                        "example": "term-111193"
                    },
                    {
                        "name": "fields",
                        "in": "query",
                        "description": "Fields available to return",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "admin.created,admin.modified,admin.id,name,summary_title",
                            "nullable": true
                        },
                        "example": "admin.id,admin.created,admin.modified,name,summary_title,description,equivalent,parent,related,identifier"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/v1/places": {
            "get": {
                "tags": [
                    "Terminology"
                ],
                "summary": "Retrieve places used in the database",
                "description": "A list of places used in the database, with pagination.",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "string"
                        },
                        "example": "Painting"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "description": "Size of the page response",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "20"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The request completed successfully."
                    },
                    "400": {
                        "description": "The request cannot be processed"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/api/v1/places/{place}": {
            "get": {
                "tags": [
                    "Terminology"
                ],
                "summary": "Retrieve a term",
                "description": "A term's details.",
                "parameters": [
                    {
                        "name": "place",
                        "in": "path",
                        "description": "Query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "string"
                        },
                        "example": "term-112039"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The request completed successfully."
                    },
                    "400": {
                        "description": "The request cannot be processed"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
...
    }
}